Moved filter_commands function to the prepare module
