+++ mysql/toolkit/script/execute.py
@@ -1,7 +1,7 @@
 from tqdm import tqdm
 from mysql.toolkit.script.dump import dump_commands, write_read_commands
 from mysql.toolkit.script.split import SplitCommands
-from mysql.toolkit.script.prepare import prepare_sql
+from mysql.toolkit.script.prepare import prepare_sql, filter_commands
 
 # Conditional import of multiprocessing module
 try:
@@ -10,21 +10,6 @@ try:
     MULTIPROCESS = True
 except ImportError:
     pass
-
-
-def filter_commands(commands, invalid_query_starts):
-    """
-    Remove particular queries from a list of SQL commands.
-
-    :param commands: List of SQL commands
-    :param invalid_query_starts: Type of SQL command to remove
-    :return: Filtered list of SQL commands
-    """
-    commands_with_drops = len(commands)
-    filtered_commands = [c for c in commands if not c.startswith(invalid_query_starts)]
-    if commands_with_drops - len(filtered_commands) > 0:
-        print("\t" + str(invalid_query_starts) + " commands removed", commands_with_drops - len(filtered_commands))
-    return filtered_commands
 
 
 class SQLScript:
